requirements.txt
numpy>=1.21.0
apify-client==1.4.1
web3==6.11.3
python-dotenv==1.0.0
discord.py==2.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
requests==2.31.0
schedule==1.2.1
python-binance==1.0.19
nltk==3.8.1
